python_folder/requirements.txt
numpy==1.16.3
optional.py==0.4.2